third_party/blink/web_tests/platform/generic/traversal/stay-within-root-expected.txt

This test checks that the NodeIterator and TreeWalker functions do not go past the root element that they are supposed to stay within.
NodeIterator: testing nextNode() (5 should be null)
[object HTMLDivElement]
[object HTMLDivElement]
[object HTMLDivElement]
null
NodeIterator: testing previousNode() (7 should be null)
null
TreeWalker: testing nextNode() (11 should be null)
[object HTMLDivElement]
[object HTMLDivElement]
null
TreeWalker: testing previousNode() (13 should be null)
null
